requirements.txt
fastapi
uvicorn
playwright
pymongo
python-dotenv
pandas
scikit-learn
numpy
streamlit
joblib